reporter: add GitHub support
Add GitHub reporter support. All of the problems in the summary are
submitted as (multi-)line comments on a given pull request. Add some
small tests to try out this functionality. I have also done some adhoc
tests.

Also, exit with 1 when there are problems with severity Bug or higher.
This is so that it would be possible to run `pint ci` in Jenkins so that
it would report problems on GitHub & exit with 1 to indicate that a step
has failed. All of the users which depend on the exit code being 0 can
simply ignore the errors with `||:` in Bash.

Signed-off-by: Giedrius Statkevičius <giedrius.statkevicius@vinted.com>