checks: series: add ignore_recordingrules option
Add a new option `ignore_recordingrules` to the series check that will
make it ignore errors when checking for the existence of metrics iff
they are part of the same pull/merge request.

This improves the series check for the CI use-case significantly because
pull/merge requests might add new recording rules together with alerting
rules, and without this option the series check will fail when they
won't be found. With this option, pint would look at the newly added
recorded rules as well.