OBS-851 ignore problems on unmodified lines
When running in CI mode all modified files are parsed and scanned. But when reporting errors to BitBucket we only send problems from modified lines (since it doesn't accept annotations on unmodified lines). This means that we can find a BUG on unmodified line and exit with non-zero code, but send a PASS report to BB.
Fix it by ignoring problems unless they are from modified lines during scan. Scan will discard anything and not pass it further if it's not modified.