fix: git changes filtering of merges
While the '--no-merges' flag excludes merge commits themselves from being included,
the commits that they introduced are still included.

This means that -- in Git workflows where people are often merging the latest changes from the main branch into theirs --
the changes from the upstream get included and evaluated, rather than just
the files changed within the pull request.

By adding the '--first-parent' flag, it ensures that only commits that have a "first parent" of the current branch will be included.

So, if you merge BranchB into BranchA, and now BranchB's commits will be included when merging BranchA to main, BranchB's commits will still be included.

However, when merging main to BranchA and then merging BranchA to main, the commits that already came from main will be ignored.

I hope this makes sense, but it's hard to explain anything to do with Git in a simple manner lol.