Fix nil pointer exceptions in certain broken dependency scenarios
When there are one or more broken dependency problems, the final
value of `dep` when evaluating all entries is the one used when
preparing the problem item.

However, it is possible that after a problem is identified,
subsequent entries are recording or alerting rules which either
do not have any vector selectors, or none that match.

In this situation, the final value of `dep` is `nil`, which causes
a crash.

This change fixes this by using the first broken dependency entry
recorded for building this part of the string.